requirements.txt
# Web framework
flask==2.3.3
requests==2.31.0

# Environment
python-dotenv==1.0.0

# Database
supabase==1.0.4

# AI
openai==1.3.8
anthropic==0.7.7

# Telegram
pyTelegramBotAPI==4.14.0

# Utilities
python-dateutil==2.8.2

google-generativeai>=0.7.0